backend/requirements.txt

# Enhanced requirements.txt
# Copy this ENTIRE content to: backend/requirements.txt

fastapi
uvicorn
openai
python-dotenv
supabase
google-generativeai
pandas
bcrypt

# Enhanced web scraping and data collection
playwright
aiohttp
beautifulsoup4
lxml
requests

# Domain and security analysis
python-whois
dnspython
pyopenssl

# Data processing and analysis
numpy
python-dateutil

# Optional: For production deployment
gunicorn
redis

# For SSL/TLS analysis
cryptography

# For async operations
asyncio-throttle
# Core API framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0

# Database
supabase==2.0.2

# AI/ML
google-generativeai==0.3.1

# Web scraping (for your scrapers)
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Data processing
pandas==2.1.3
numpy==1.24.3

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Utilities
pydantic==2.5.0
typing-extensions==4.8.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
supabase==2.0.2
pydantic==2.5.0